use serde remote attribute when deriving Deserialize
i was so excited last night about a potential solution i
had found. see my last commit message for the details.
today i woke up ready to implement all the things! to then
realize that rust does not allow you to implement a trait
for a type multiple times. never ever. trait
implementations are not scoped, and they cannot be
overridden.

as someone who is still very new to rust, this is one of
quite a few limitations that have had me scratching my
head. maybe one day i'll get it.

anyways i've reverted back to using an attribute called
remote when deriving the serde Deserialize trait. it's
designed for implementing Deserialize on types from other
crates, so works OK for this use-case of implementing it
on a type from a different module.

the big downside of using this attribute (and, why i didn't
do it originally) is that it does not play nicely with
wrapped values in a struct or enum, like Vec<MyType>. you
have to hold serde's hand in these cases, which is a real
pain and starts to defeat the purpose of all this macro
magic. i am borrowing this example code [0] from an issue
in the serde repository. hopefully, serde will support this
case better soon.