create a Reader trait that each reader implements
this is a step closer to some sort of standard API for the
readers. i originally wanted to do this so that each reader could
handle its own initialization data, such as a buffered reader or
file path.

if this works well for the readers, i'll consider doing something
similar for the writers.